Move code into Game component
